+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fixed incorrect label length calculation.
+-   Fixed default user Build Actions path for new installations (should be `Assets/SuperUnityBuild/Editor/BuildActions`).
 
 ## [5.0.1] - 2021-11-25